Abstimmung im Roboter, Knopf auf der Weboberflaeche
Die Abstimmung laeuft jetzt auch ohne PC: ein nicht-blockierender
Zustandsautomat, der aus dem 200-Hz-Regeltakt getickt wird und ihn nie
aufhaelt. Dieselben Stufen und dasselbe Twiddle wie autotune.py, nur im
Roboter selbst - der Knopf auf der Seite kann kein PC-Skript starten.

Faellt er waehrend eines Versuchs um, zaehlt der als schlechtestes
Ergebnis; der Automat wartet dann, bis jemand ihn wieder aufstellt, und
macht von selbst weiter. STOP und ein zweiter Druck auf den Knopf
beenden mit den besten gefundenen Werten, TUNEOFF verwirft sie.

minPwm ist dabei auf float umgestellt, damit der Abstimmer alle
Parameter einheitlich anfassen kann (NVS-Schluessel minpwm2).

+++ include/state.h
@@ -11,7 +11,7 @@ extern bool  fallFlag;
 extern int   lastPwmOut;
 
 extern float Kp, Ki, Kd, trim, outSign;
-extern int   minPwm;
+extern float minPwm;
 
 // Gierregler: haelt die Blickrichtung ueber Gyro-Z
 extern float yawDeg;
